Fix for: A9:2013 / A6:2021 - Vuln and outdated components
incremented the marked version to latest and
used correct semvar value ^ to keep it updated

apart from that accomodated how marked is now used,
it needs an external output sanitizer so I added it.
The recommended output sanitizer requires a dom emulation lib to work
server-side, so I added jsdom.